Исправила исключение
